ConverterTextToCollection extracted ConverterTextToList ConverterTextToCollectionCollectionList
